PPF-25: fill method exceptions

+++ PyPDFForm/wrapper.py
@@ -22,7 +22,7 @@ class PyPDFForm(object):
         return new_obj
 
     def fill(
-        self, data: dict, font_size: float = 12, text_wrap_length: int = 100,
+        self, data: dict, font_size: int = 12, text_wrap_length: int = 100,
     ) -> "PyPDFForm":
         self.stream = (
             _PyPDFForm()
